Revert "feat(rosettarpc): Data API implementation (#2940)"
This reverts commit 4e896171a60c2d7ee232aca6bc44cf20afe1dc62.
