+++ Cargo.toml
@@ -30,7 +30,6 @@ members = [
     "chain/jsonrpc",
     "chain/jsonrpc/client",
     "chain/jsonrpc/test-utils",
-    "chain/rosetta-rpc",
     "test-utils/testlib",
     "test-utils/loadtester",
     "test-utils/state-viewer",
@@ -105,4 +104,3 @@ adversarial = ["neard/adversarial", "near-jsonrpc/adversarial", "near-store/adve
 no_cache = ["neard/no_cache"]
 metric_recorder = ["neard/metric_recorder"]
 delay_detector = ["neard/delay_detector"]
-rosetta_rpc = ["neard/rosetta_rpc"]
