--- chain/rosetta-rpc/CHANGELOG.md
@@ -1,6 +1,0 @@
-# Changelog
-
-## 0.1.0
-
-* Data API exposes feature-complete balance-changing operations
-* Construction API exposes the bare minimum signed transaction submittion API
