--- chain/rosetta-rpc/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "near-rosetta-rpc"
-version = "0.1.0"
-authors = ["Near Inc <hello@nearprotocol.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lazy_static = "1.4"
-actix = "0.9"
-actix-web = "2"
-actix-cors = "0.2"
-futures = "0.3.5"
-tokio = { version = "0.2", features = ["full"] }
-paperclip = { version = "0.4", features = ["actix", "actix-nightly"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-strum = { version = "0.18", features = ["derive"] }
-validator = "0.10"
-
-near-primitives = { path = "../../core/primitives" }
-near-crypto = { path = "../../core/crypto" }
-near-chain-configs = { path = "../../core/chain-configs" }
-near-client = { path = "../client" }
-near-network = { path = "../network" }
-near-runtime-configs = { path = "../../core/runtime-configs" }
