--- chain/rosetta-rpc/rosetta.cfg
@@ -1,92 +1,0 @@
-{
- "network": {
-  "blockchain": "nearprotocol",
-  "network": "testnet"
- },
- "online_url": "http://localhost:3040",
- "data_directory": "",
- "http_timeout": 10,
- "retry_elapsed_time": 60,
- "sync_concurrency": 8,
- "transaction_concurrency": 16,
- "tip_delay": 300,
- "disable_memory_limit": false,
- "log_configuration": false,
- "construction": {
-  "offline_url": "http://localhost:3040",
-  "currency": {
-   "symbol": "yoctoNEAR",
-   "decimals": 0
-  },
-  "minimum_balance": "0",
-  "maximum_fee": "5000000000000000",
-  "curve_type": "secp256k1",
-  "accounting_model": "account",
-  "scenario": [
-   {
-    "operation_identifier": {
-     "index": 0
-    },
-    "type": "transfer",
-    "status": "",
-    "account": {
-     "address": "{{ SENDER }}"
-    },
-    "amount": {
-     "value": "{{ SENDER_VALUE }}",
-     "currency": null
-    }
-   },
-   {
-    "operation_identifier": {
-     "index": 1
-    },
-    "related_operations": [
-     {
-      "index": 0
-     }
-    ],
-    "type": "transfer",
-    "status": "",
-    "account": {
-     "address": "{{ RECIPIENT }}"
-    },
-    "amount": {
-     "value": "{{ RECIPIENT_VALUE }}",
-     "currency": null
-    }
-   }
-  ],
-  "confirmation_depth": 10,
-  "stale_depth": 30,
-  "broadcast_limit": 3,
-  "ignore_broadcast_failures": false,
-  "change_scenario": null,
-  "clear_broadcasts": false,
-  "broadcast_behind_tip": false,
-  "block_broadcast_limit": 5,
-  "rebroadcast_all": false,
-  "new_account_probability": 0.5,
-  "max_addresses": 200
- },
- "data": {
-  "start_index": 12140933,
-  "active_reconciliation_concurrency": 16,
-  "inactive_reconciliation_concurrency": 4,
-  "inactive_reconciliation_frequency": 250,
-  "log_blocks": false,
-  "log_transactions": false,
-  "log_balance_changes": false,
-  "log_reconciliations": false,
-  "ignore_reconciliation_error": false,
-  "exempt_accounts": "",
-  "bootstrap_balances": "",
-  "historical_balance_disabled": false,
-  "interesting_accounts": "",
-  "reconciliation_disabled": false,
-  "inactive_discrepency_search_disabled": false,
-  "balance_tracking_disabled": false,
-  "coin_tracking_disabled": false,
-  "results_output_file": ""
- }
-}
