+++ deny.toml
@@ -11,9 +11,6 @@ deny = [
 skip = [
     # actix 0.9.0 still uses it
     { name = "tokio-util", version = "=0.2.0" },
-    { name = "parking_lot", version = "=0.10.2" },
-    { name = "parking_lot_core", version = "=0.7.2" },
-    { name = "lock_api", version = "=0.3.4" },
     # actix-server 1.0.2 still uses it
     { name = "miow", version = "=0.2.1" },
     # miow 0.6.21 still uses it
@@ -37,10 +34,4 @@ skip = [
     { name = "wasmparser", version = "=0.51.4"},
     { name = "itertools", version = "=0.8.2" },
     { name = "rand_core", version = "=0.4.2" },
-
-    # rosetta-rpc pull paperclip which introduce the following duplocates (https://github.com/wafflespeanut/paperclip/pull/209)
-    { name = "url", version = "=1.7.2" },
-    { name = "idna", version = "=0.1.5" },
-    { name = "semver", version = "=0.9.0" },
-    { name = "percent-encoding", version = "=1.0.1" },
 ]
