+++ neard/src/config.rs
@@ -31,8 +31,6 @@ use near_primitives::types::{
 use near_primitives::utils::{generate_random_string, get_num_seats_per_shard};
 use near_primitives::validator_signer::{InMemoryValidatorSigner, ValidatorSigner};
 use near_primitives::version::PROTOCOL_VERSION;
-#[cfg(feature = "rosetta_rpc")]
-use near_rosetta_rpc::RosettaRpcConfig;
 use near_runtime_configs::RuntimeConfig;
 use near_telemetry::TelemetryConfig;
 
@@ -384,8 +382,6 @@ pub struct Config {
     pub validator_key_file: String,
     pub node_key_file: String,
     pub rpc: RpcConfig,
-    #[cfg(feature = "rosetta_rpc")]
-    pub rosetta_rpc: RosettaRpcConfig,
     pub telemetry: TelemetryConfig,
     pub network: Network,
     pub consensus: Consensus,
@@ -406,8 +402,6 @@ impl Default for Config {
             validator_key_file: VALIDATOR_KEY_FILE.to_string(),
             node_key_file: NODE_KEY_FILE.to_string(),
             rpc: RpcConfig::default(),
-            #[cfg(feature = "rosetta_rpc")]
-            rosetta_rpc: RosettaRpcConfig::default(),
             telemetry: TelemetryConfig::default(),
             network: Network::default(),
             consensus: Consensus::default(),
@@ -529,8 +523,6 @@ pub struct NearConfig {
     pub client_config: ClientConfig,
     pub network_config: NetworkConfig,
     pub rpc_config: RpcConfig,
-    #[cfg(feature = "rosetta_rpc")]
-    pub rosetta_rpc_config: RosettaRpcConfig,
     pub telemetry_config: TelemetryConfig,
     pub genesis: Arc<Genesis>,
     pub validator_signer: Option<Arc<dyn ValidatorSigner>>,
@@ -626,8 +618,6 @@ impl NearConfig {
             },
             telemetry_config: config.telemetry,
             rpc_config: config.rpc,
-            #[cfg(feature = "rosetta_rpc")]
-            rosetta_rpc_config: config.rosetta_rpc,
             genesis,
             validator_signer,
         }
