+++ runtime/runtime/src/verifier.rs
@@ -1,3 +1,6 @@
+use crate::actions::get_insufficient_storage_stake;
+use crate::config::{total_prepaid_gas, tx_cost, RuntimeConfig, TransactionCost};
+use crate::VerificationResult;
 use near_crypto::key_conversion::is_valid_staking_key;
 use near_primitives::account::AccessKeyPermission;
 use near_primitives::errors::{
@@ -10,15 +13,9 @@ use near_primitives::transaction::{
     SignedTransaction, StakeAction,
 };
 use near_primitives::utils::is_valid_account_id;
-use near_runtime_configs::get_insufficient_storage_stake;
-use near_store::{
-    get_access_key, get_account, set_access_key, set_account, StorageError, TrieUpdate,
-};
+use near_store::{get_access_key, get_account, set_access_key, set_account, TrieUpdate};
 use near_vm_logic::types::Balance;
 use near_vm_logic::VMLimitConfig;
-
-use crate::config::{total_prepaid_gas, tx_cost, RuntimeConfig, TransactionCost};
-use crate::VerificationResult;
 
 /// Validates the transaction without using the state. It allows any node to validate a
 /// transaction before forwarding it to the node that tracks the `signer_id` account.
@@ -131,9 +128,7 @@ pub fn verify_and_charge_transaction(
             }
             .into())
         }
-        Err(err) => {
-            return Err(RuntimeError::StorageError(StorageError::StorageInconsistentState(err)))
-        }
+        Err(err) => return Err(RuntimeError::StorageError(err)),
     };
 
     if let AccessKeyPermission::FunctionCall(ref function_call_permission) = access_key.permission {
